+++ src/main/java/Constants/Emoji.java
@@ -112,6 +112,7 @@ public class Emoji {
     public final static String GLOBE = EmojiParser.parseToUnicode(":globe_with_meridians:");
     public final static String NOTES = EmojiParser.parseToUnicode(":notes:");
     public final static String RADIO = EmojiParser.parseToUnicode(":radio_button:");
+    public final static String AUTOPLAY = EmojiParser.parseToUnicode(":leftwards_arrow_with_hook:");
     public final static String PAUSE = EmojiParser.parseToUnicode(":double_vertical_bar:");
     public final static String NEXT_TRACK = EmojiParser.parseToUnicode(":black_right_pointing_double_triangle_with_vertical_bar:");
     public final static String UP_VOTE = EmojiParser.parseToUnicode(":arrow_up_small:");